+++ src/safeds/ml/nn/_model.py
@@ -20,7 +20,7 @@ from safeds.ml.nn.converters import (
     OutputConversionImageToImage,
     OutputConversionImageToTable,
 )
-from safeds.ml.nn.converters._output_conversion_image import _OutputConversionImage
+from safeds.ml.nn.converters._output_converter_image import _OutputConversionImage
 from safeds.ml.nn.layers import (
     Convolutional2DLayer,
     FlattenLayer,
