+++ src/safeds/ml/nn/converters/_output_converter.py
