+++ src/safeds/ml/nn/converters/_output_converter_time_series.py
@@ -7,7 +7,7 @@ from safeds._utils import _structural_hash
 from safeds.data.labeled.containers import TimeSeriesDataset
 from safeds.data.tabular.containers import Column
 
-from ._output_conversion import OutputConversion
+from ._output_converter import OutputConversion
 
 if TYPE_CHECKING:
     from torch import Tensor
