+++ tests/safeds/ml/nn/converters/test_input_converter_image.py
